Finished tests for helper file

+++ studentdb/tabledb.go
@@ -16,7 +16,7 @@ func CreateClassTable(db *sql.DB) error {
 }
 
 // InsertStudent - Creates a new student into the class_records table
-func InsertStudent(db *sql.DB, name string, grade float32) (int, error) {
+func InsertStudent(db *sql.DB, name string, grade float64) (int, error) {
 	statement := `INSERT INTO class_records(name,average) VALUES($1, $2) RETURNING id`
 	var id int
 	err := db.QueryRow(statement, name, grade).Scan(&id)
@@ -27,7 +27,7 @@ func InsertStudent(db *sql.DB, name string, grade float32) (int, error) {
 }
 
 // UpdateUser - Update a student grade by there id
-func UpdateUser(db *sql.DB, id int, grade float32) error {
+func UpdateUser(db *sql.DB, id int, grade float64) error {
 	statement := `UPDATE class_records SET average=$2 WHERE id=$1`
 	_, err := db.Exec(statement, id, grade)
 	return err
